[fix] sale price can be float

+++ src/lib/web3connect.ts
@@ -260,7 +260,7 @@ export class Web3Connect {
         if (sale) {
           sale = {
             end: parseInt(sale.end.toString()),
-            price: parseInt(Web3.utils.fromWei(sale.price.toString())),
+            price: parseFloat(Web3.utils.fromWei(sale.price.toString())),
           }
         }
       }
